trading-agent/requirements.txt
nats-py>=2.6.0
asyncpg>=0.29.0
httpx>=0.26.0
cryptography>=42.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=24.1.0
anthropic>=0.39.0
